+++ src/main/java/seedu/ibook/ui/CommandBox.java
@@ -5,6 +5,9 @@ import javafx.scene.control.TextField;
 import javafx.scene.layout.HBox;
 import seedu.ibook.ui.popup.PopupAdd;
 
+/**
+ * CommandBox Ui class.
+ */
 public class CommandBox extends UiPart<HBox> {
 
     private static final String FXML = "CommandBox.fxml";
@@ -16,7 +19,11 @@ public class CommandBox extends UiPart<HBox> {
     private TextField commandTextField;
 
     /**
-     * Creates a {@code CommandBox} with the given {@code CommandExecutor}.
+     * Creates a {@code CommandBox} with a {@code CommandExecutor}
+     * and a {@code popupAdd}.
+     *
+     * @param commandExecutor Function to execute command.
+     * @param popupAdd PopupAdd window to create product.
      */
     public CommandBox(MainWindow.CommandExecutor commandExecutor,
                       PopupAdd popupAdd) {
@@ -26,7 +33,7 @@ public class CommandBox extends UiPart<HBox> {
     }
 
     /**
-     * Handles the Enter button pressed event.
+     * Handles the Enter button pressed event in the command line.
      */
     @FXML
     private void handleCommandEntered() {
@@ -39,8 +46,11 @@ public class CommandBox extends UiPart<HBox> {
         commandTextField.setText("");
     }
 
+    /**
+     * Handles the add product button clicked event.
+     */
     @FXML
-    private void handlePopupAdd() {
+    private void handleAddProductClicked() {
         if (popupAdd.isShowing()) {
             popupAdd.focus();
         } else {
