+++ src/main/java/seedu/ibook/ui/ResultWindow.java
@@ -2,23 +2,29 @@ package seedu.ibook.ui;
 
 import static java.util.Objects.requireNonNull;
 
-import javafx.fxml.FXML;
 import javafx.scene.control.TextArea;
 
+/**
+ * The window that shows feedback of a command.
+ */
 public class ResultWindow extends UiPart<TextArea> {
 
     private static final String FXML = "ResultWindow.fxml";
 
-    @FXML
-    private TextArea resultDisplay;
-
+    /**
+     * Initializes a {@code ResultWindow}.
+     */
     ResultWindow() {
         super(FXML);
     }
 
+    /**
+     * Shows the feedback of a command to user.
+     * @param feedbackToUser The feedback of a command.
+     */
     void setFeedbackToUser(String feedbackToUser) {
         requireNonNull(feedbackToUser);
-        resultDisplay.setText(feedbackToUser);
+        getRoot().setText(feedbackToUser);
     }
 
 }
