--- src/main/java/seedu/ibook/ui/TableHeader.java
@@ -1,12 +1,0 @@
-package seedu.ibook.ui;
-
-import javafx.scene.layout.HBox;
-
-public class TableHeader extends UiPart<HBox> {
-
-    private static final String FXML = "TableHeader.fxml";
-
-    TableHeader() {
-        super(FXML);
-    }
-}
