+++ src/main/java/seedu/ibook/ui/table/Table.java
@@ -1,15 +1,19 @@
-package seedu.ibook.ui;
+package seedu.ibook.ui.table;
 
 import javafx.collections.ObservableList;
 import javafx.scene.Node;
 import javafx.scene.layout.VBox;
 import seedu.ibook.model.product.Product;
+import seedu.ibook.ui.UiPart;
 import seedu.ibook.ui.popup.PopupDelete;
 import seedu.ibook.ui.popup.PopupUpdate;
 
+/**
+ * The table that is containing {@code TableContent} and {@code TableHeader}
+ */
 public class Table extends UiPart<VBox> {
 
-    private static final String FXML = "Table.fxml";
+    private static final String FXML = "Table/Table.fxml";
 
     private TableHeader tableHeader;
     private TableContent tableContent;
@@ -19,7 +23,14 @@ public class Table extends UiPart<VBox> {
 
     private final ObservableList<Product> filteredIBook;
 
-    Table(ObservableList<Product> filteredIBook,
+    /**
+     * Initializes a {@code Table}.
+     *
+     * @param filteredIBook The filtered {@code Product} list.
+     * @param popupUpdate The popup responsible for update.
+     * @param popupDelete The popup responsible for delete.
+     */
+    public Table(ObservableList<Product> filteredIBook,
           PopupUpdate popupUpdate,
           PopupDelete popupDelete) {
         super(FXML);
@@ -29,7 +40,7 @@ public class Table extends UiPart<VBox> {
         populateField();
     }
 
-    void populateField() {
+    private void populateField() {
         ObservableList<Node> children = getRoot().getChildren();
 
         tableHeader = new TableHeader();
