+++ src/main/java/seedu/ibook/ui/table/TableContent.java
@@ -1,4 +1,4 @@
-package seedu.ibook.ui;
+package seedu.ibook.ui.table;
 
 import javafx.collections.ListChangeListener;
 import javafx.collections.ObservableList;
@@ -6,12 +6,16 @@ import javafx.fxml.FXML;
 import javafx.scene.control.ScrollPane;
 import javafx.scene.layout.VBox;
 import seedu.ibook.model.product.Product;
+import seedu.ibook.ui.UiPart;
 import seedu.ibook.ui.popup.PopupDelete;
 import seedu.ibook.ui.popup.PopupUpdate;
 
+/**
+ * The main content of the table.
+ */
 public class TableContent extends UiPart<ScrollPane> {
 
-    private static final String FXML = "TableContent.fxml";
+    private static final String FXML = "Table/TableContent.fxml";
 
     private final ObservableList<Product> filteredIBook;
 
@@ -21,6 +25,13 @@ public class TableContent extends UiPart<ScrollPane> {
     @FXML
     private VBox content;
 
+    /**
+     * Initializes a {@code TableContent}.
+     *
+     * @param filteredIBook The filtered {@code Product} list.
+     * @param popupUpdate The popup responsible for update.
+     * @param popupDelete The popup responsible for delete.
+     */
     TableContent(ObservableList<Product> filteredIBook,
                  PopupUpdate popupUpdate,
                  PopupDelete popupDelete) {
@@ -32,7 +43,7 @@ public class TableContent extends UiPart<ScrollPane> {
         populateField();
     }
 
-    void populateField() {
+    private void populateField() {
         content.getChildren().clear();
         for (int i = 0; i < filteredIBook.size(); i++) {
             Product product = filteredIBook.get(i);
