+++ src/main/resources/view/MainWindow.fxml
@@ -5,14 +5,15 @@
 <?import javafx.scene.layout.VBox?>
 <?import javafx.stage.Stage?>
 
-<fx:root type="javafx.stage.Stage" title="iBook" xmlns="http://javafx.com/javafx/8"
-         xmlns:fx="http://javafx.com/fxml/1" minHeight="700.0" minWidth="900.0">
+<fx:root type="javafx.stage.Stage"
+         xmlns="http://javafx.com/javafx/11" xmlns:fx="http://javafx.com/fxml/1"
+         title="iBook"  minHeight="700.0" minWidth="900.0">
     <scene>
-        <Scene stylesheets="@global.css">
-            <AnchorPane minHeight="700.0" minWidth="900.0" prefHeight="700.0" prefWidth="900.0"
-                        style="-fx-background-color: #FFFFFF;">
-                <VBox fx:id="mainContent" AnchorPane.bottomAnchor="0.0" AnchorPane.leftAnchor="0.0"
-                      AnchorPane.rightAnchor="0.0" AnchorPane.topAnchor="0.0">
+        <Scene stylesheets="@css/global.css">
+            <AnchorPane styleClass="main">
+                <VBox fx:id="mainContent" AnchorPane.bottomAnchor="0.0"
+                      AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0"
+                      AnchorPane.topAnchor="0.0">
                 </VBox>
             </AnchorPane>
         </Scene>
