--- src/main/resources/view/ProductCard.fxml
@@ -1,77 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.geometry.Insets?>
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.control.Label?>
-<?import javafx.scene.image.Image?>
-<?import javafx.scene.image.ImageView?>
-<?import javafx.scene.layout.HBox?>
-
-<HBox xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1"
-      maxHeight="1.7976931348623157E308" maxWidth="1.7976931348623157E308" spacing="5.0"
-      style="-fx-border-style: solid; -fx-border-width: 0 0 0.1 0;">
-    <padding>
-        <Insets bottom="10.0" left="10.0" right="10.0" top="10.0"/>
-    </padding>
-    <Label fx:id="indexLabel" contentDisplay="CENTER" maxHeight="1.7976931348623157E308"
-           maxWidth="1.7976931348623157E308" prefWidth="70.0">
-        <padding>
-            <Insets left="10.0" right="10.0"/>
-        </padding>
-    </Label>
-    <Label fx:id="name" contentDisplay="CENTER" maxHeight="1.7976931348623157E308"
-           maxWidth="1.7976931348623157E308" textOverrun="CLIP" wrapText="true"
-           prefWidth="100.0">
-        <padding>
-            <Insets left="10.0" right="10.0"/>
-        </padding>
-    </Label>
-    <Label fx:id="category" contentDisplay="CENTER" maxHeight="1.7976931348623157E308"
-           maxWidth="1.7976931348623157E308" textOverrun="CLIP" wrapText="true"
-           prefWidth="100.0">
-        <padding>
-            <Insets left="10.0" right="10.0"/>
-        </padding>
-    </Label>
-    <Label fx:id="expiryDate" contentDisplay="CENTER" maxHeight="1.7976931348623157E308"
-           maxWidth="1.7976931348623157E308" textOverrun="CLIP" wrapText="true"
-           prefWidth="100.0">
-        <padding>
-            <Insets left="10.0" right="10.0"/>
-        </padding>
-    </Label>
-    <Label fx:id="price" contentDisplay="CENTER" maxHeight="1.7976931348623157E308"
-           maxWidth="1.7976931348623157E308" textOverrun="CLIP" wrapText="true"
-           prefWidth="100.0">
-        <padding>
-            <Insets left="10.0" right="10.0"/>
-        </padding>
-    </Label>
-    <Label fx:id="description" contentDisplay="CENTER" maxHeight="1.7976931348623157E308"
-           maxWidth="1.7976931348623157E308" prefWidth="100.0" textOverrun="CLIP" wrapText="true"
-           HBox.hgrow="ALWAYS">
-        <padding>
-            <Insets left="10.0" right="10.0"/>
-        </padding>
-    </Label>
-    <HBox fx:id="action" alignment="CENTER" maxHeight="1.7976931348623157E308"
-          maxWidth="1.7976931348623157E308" prefWidth="100.0">
-        <padding>
-            <Insets left="10.0"/>
-        </padding>
-        <Button alignment="CENTER" onMouseClicked="#handlePopupUpdate" mnemonicParsing="false" styleClass="edit-button">
-            <graphic>
-                <ImageView fitHeight="20.0" fitWidth="20.0" pickOnBounds="true" preserveRatio="true">
-                    <Image url="@../images/edit-color.png"/>
-                </ImageView>
-            </graphic>
-        </Button>
-        <Button alignment="CENTER" onMouseClicked="#handlePopupDelete" mnemonicParsing="false" styleClass="delete-button">
-            <graphic>
-                <ImageView fitHeight="20.0" fitWidth="20.0" pickOnBounds="true" preserveRatio="true">
-                    <Image url="@../images/trash-2-color.png"/>
-                </ImageView>
-            </graphic>
-        </Button>
-    </HBox>
-</HBox>
