+++ src/main/resources/view/ResultWindow.fxml
@@ -4,14 +4,8 @@
 <?import javafx.scene.layout.VBox?>
 <?import javafx.geometry.Insets?>
 
-<TextArea fx:id="resultDisplay" style="-fx-background-color: ffffff;
-        -fx-effect: dropshadow(gaussian, rgba(0,0,0,0.2), 18, 0, 0, 5);"
-          prefHeight="100.0" editable="false" xmlns="http://javafx.com/javafx/8"
-          wrapText="true"
-          xmlns:fx="http://javafx.com/fxml/1">
-    <padding>
-        <Insets bottom="10.0" left="10.0" right="10.0" top="10.0"/>
-    </padding>
+<TextArea editable="false" wrapText="true"
+          styleClass="result-window" stylesheets="@css/resultWindow.css">
     <VBox.margin>
         <Insets bottom="20.0" left="40.0" right="40.0" top="20.0"/>
     </VBox.margin>
