+++ src/main/resources/view/Table/ProductCard.fxml
@@ -1,0 +1,33 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.control.Label?>
+<?import javafx.scene.image.Image?>
+<?import javafx.scene.image.ImageView?>
+<?import javafx.scene.layout.HBox?>
+
+<HBox xmlns="http://javafx.com/javafx/11" xmlns:fx="http://javafx.com/fxml/1"
+      styleClass="table-row">
+    <Label fx:id="indexLabel" styleClass="table-column-short"/>
+    <Label fx:id="name" styleClass="table-column-long"/>
+    <Label fx:id="category" styleClass="table-column-long"/>
+    <Label fx:id="expiryDate" styleClass="table-column-long"/>
+    <Label fx:id="price" styleClass="table-column-short"/>
+    <Label fx:id="description" styleClass="table-column-long" HBox.hgrow="ALWAYS"/>
+    <HBox fx:id="action" styleClass="table-column-long">
+        <Button alignment="CENTER" onMouseClicked="#handlePopupUpdate" mnemonicParsing="false" styleClass="edit">
+            <graphic>
+                <ImageView fitHeight="20.0" fitWidth="20.0" pickOnBounds="true" preserveRatio="true">
+                    <Image url="@../../images/edit-color.png"/>
+                </ImageView>
+            </graphic>
+        </Button>
+        <Button alignment="CENTER" onMouseClicked="#handlePopupDelete" mnemonicParsing="false" styleClass="delete">
+            <graphic>
+                <ImageView fitHeight="20.0" fitWidth="20.0" pickOnBounds="true" preserveRatio="true">
+                    <Image url="@../../images/trash-2-color.png"/>
+                </ImageView>
+            </graphic>
+        </Button>
+    </HBox>
+</HBox>
