+++ src/main/resources/view/Table/Table.fxml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.geometry.Insets?>
+<?import javafx.scene.layout.VBox?>
+
+<VBox VBox.vgrow="ALWAYS" stylesheets="@../css/table.css" styleClass="table">
+    <VBox.margin>
+        <Insets bottom="20.0" left="20.0" right="20.0" />
+    </VBox.margin>
+</VBox>
