+++ src/main/resources/view/Table/TableContent.fxml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.control.ScrollPane?>
+<?import javafx.scene.layout.VBox?>
+
+<ScrollPane xmlns="http://javafx.com/javafx/11" xmlns:fx="http://javafx.com/fxml/1"
+            fitToHeight="true" fitToWidth="true"
+            VBox.vgrow="ALWAYS" styleClass="table-content">
+    <VBox fx:id="content" style="-fx-background-color: white;"/>
+</ScrollPane>
