+++ src/main/resources/view/Table/TableHeader.fxml
@@ -1,0 +1,14 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.control.Label?>
+<?import javafx.scene.layout.HBox?>
+
+<HBox styleClass="table-header">
+    <Label styleClass="table-column-short" text="INDEX"/>
+    <Label styleClass="table-column-long" text="NAME"/>
+    <Label styleClass="table-column-long" text="CATEGORY"/>
+    <Label styleClass="table-column-long" text="EXPIRY DATE"/>
+    <Label styleClass="table-column-short" text="PRICE"/>
+    <Label styleClass="table-column-long" text="DESCRIPTION" HBox.hgrow="ALWAYS"/>
+    <Label styleClass="table-column-long" HBox.hgrow="NEVER"/>
+</HBox>
