--- src/main/resources/view/TableContent.fxml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.control.ScrollPane?>
-<?import javafx.scene.layout.VBox?>
-
-<ScrollPane xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1"
-            fitToHeight="true" fitToWidth="true" style="-fx-background-color: transparent;"
-            VBox.vgrow="ALWAYS">
-    <VBox fx:id="content" style="-fx-background-color: #ffffff;"/>
-</ScrollPane>
