--- src/main/resources/view/TableHeader.fxml
@@ -1,50 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.geometry.Insets?>
-<?import javafx.scene.control.Label?>
-<?import javafx.scene.layout.HBox?>
-
-<HBox spacing="5.0" style="-fx-border-style: solid; -fx-border-width: 0 0 0.1 0;">
-    <padding>
-        <Insets bottom="10.0" left="10.0" right="10.0" top="10.0"/>
-    </padding>
-    <Label contentDisplay="CENTER" prefWidth="70.0" styleClass="grid-title" text="INDEX">
-        <padding>
-            <Insets left="10.0" right="10.0"/>
-        </padding>
-    </Label>
-    <Label contentDisplay="CENTER" prefWidth="100.0" styleClass="grid-title" text="NAME">
-        <HBox.margin>
-            <Insets/>
-        </HBox.margin>
-        <padding>
-            <Insets left="10.0" right="10.0"/>
-        </padding>
-    </Label>
-    <Label contentDisplay="CENTER" prefWidth="100.0" styleClass="grid-title" text="CATEGORY">
-        <padding>
-            <Insets left="10.0" right="10.0"/>
-        </padding>
-    </Label>
-    <Label contentDisplay="CENTER" prefWidth="100.0" styleClass="grid-title" text="EXPIRY DATE">
-        <padding>
-            <Insets left="10.0" right="10.0"/>
-        </padding>
-    </Label>
-    <Label contentDisplay="CENTER" prefWidth="100.0" styleClass="grid-title" text="PRICE">
-        <padding>
-            <Insets left="10.0" right="10.0"/>
-        </padding>
-    </Label>
-    <Label contentDisplay="CENTER" maxWidth="1.7976931348623157E308" prefWidth="100.0" styleClass="grid-title"
-           text="DESCRIPTION" HBox.hgrow="ALWAYS">
-        <padding>
-            <Insets left="10.0" right="10.0"/>
-        </padding>
-    </Label>
-    <Label contentDisplay="CENTER" prefWidth="100.0" styleClass="grid-title" HBox.hgrow="NEVER">
-        <padding>
-            <Insets left="10.0"/>
-        </padding>
-    </Label>
-</HBox>
