+++ src/main/resources/view/css/resultWindow.css
@@ -1,0 +1,16 @@
+.result-window {
+    -fx-pref-height: 100;
+    -fx-background-color: white;
+    -fx-effect: dropshadow(gaussian, rgba(0,0,0,0.2), 18, 0, 0, 5);
+    -fx-padding: 10;
+    -fx-font-size: 16;
+}
+
+.text-area {
+    -fx-background-insets: 0;
+    -fx-background-color: white;
+}
+
+.text-area .content {
+    -fx-background-color: white;
+}
