--- src/main/resources/view/global.css
@@ -1,99 +1,0 @@
-@font-face {
-    src: url('font/NotoSans-Regular.ttf');
-}
-
-* {
-    -fx-font-family: 'Noto Sans';
-}
-
-.grid-title {
-    -fx-border-width: 0 0 0 0.1;
-    -fx-border-style: solid;
-    -fx-text-fill: #AEAEAE;
-}
-
-.menu .label {
-    -fx-text-fill: #FFFFFF;
-}
-
-.menu:hover,
-.menu:showing {
-    -fx-background-color: #8373e7;
-    -fx-text-fill: #FFFFFF;
-}
-
-.menu-item {
-    -fx-background-color: #FFFFFF;
-    -fx-text-fill: #000000;
-}
-
-.menu-item .label {
-    -fx-text-fill: #000000;
-}
-
-.menu-item:focused .label {
-    -fx-text-fill: #000000;
-}
-
-.menu-item:focused {
-    -fx-background: -fx-accent;
-    -fx-background-color: #E0E0E0;
-}
-
-.command {
-    -fx-background-color: #F1F2F3;
-    -fx-background-radius: 8px;
-    -fx-padding: 15;
-}
-
-.command:focused {
-    -fx-background-color: white;
-    -fx-effect: dropshadow(gaussian, rgba(0, 0, 0, 0.2), 3.5, 0.4, 0, 1);
-    -fx-prompt-text-fill: derive(-fx-control-inner-background,-30%);
-}
-
-.text-area {
-    -fx-background-insets: 0;
-    -fx-background-color: white;
-}
-
-.text-area .content {
-    -fx-background-color: white;
-}
-
-.primary-button {
-    -fx-background-color: #9F93Ed;
-    -fx-background-radius: 8;
-}
-
-.primary-button:hover {
-    -fx-background-color: #8373e7;
-}
-
-.primary-button:pressed {
-    -fx-background-color: #7d6de0;
-}
-
-.edit-button {
-    -fx-background-color: transparent;
-}
-
-.edit-button:hover {
-    -fx-background-color: #f4f5f6;
-}
-
-.edit-button:pressed {
-    -fx-background-color: #f1f2f3;
-}
-
-.delete-button {
-    -fx-background-color: transparent;
-}
-
-.delete-button:hover {
-    -fx-background-color: #f4f5f6;
-}
-
-.delete-button:pressed {
-    -fx-background-color: #f1f2f3;
-}
