+++ src/main/resources/view/popup/PopupAdd.fxml
@@ -11,25 +11,21 @@
 <?import javafx.scene.layout.AnchorPane?>
 <?import javafx.scene.layout.HBox?>
 <?import javafx.scene.layout.VBox?>
-<?import javafx.scene.text.Font?>
 <?import javafx.scene.text.Text?>
 <?import javafx.stage.Stage?>
 
-<fx:root resizable="false" title="Add Product" type="javafx.stage.Stage" xmlns="http://javafx.com/javafx/11" xmlns:fx="http://javafx.com/fxml/1">
+<fx:root xmlns="http://javafx.com/javafx/11" xmlns:fx="http://javafx.com/fxml/1"
+         type="javafx.stage.Stage" resizable="false" title="Add Product" >
     <scene>
-        <Scene stylesheets="@popup.css">
-           <AnchorPane style="-fx-background-color: #FFFFFF;">
-              <VBox AnchorPane.bottomAnchor="0.0" AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0"
-                    AnchorPane.topAnchor="0.0">
+        <Scene stylesheets="@../css/popup.css">
+           <AnchorPane style="-fx-background-color: white;">
+              <VBox AnchorPane.bottomAnchor="0.0" AnchorPane.leftAnchor="0.0"
+                    AnchorPane.rightAnchor="0.0" AnchorPane.topAnchor="0.0">
                  <padding>
                     <Insets bottom="20.0" left="20.0" right="20.0" top="20.0"/>
                  </padding>
-                 <Label text="Product" styleClass="title" minWidth="100">
-                    <font>
-                       <Font size="28.0"/>
-                    </font>
-                 </Label>
-                 <HBox maxWidth="1.7976931348623157E308" spacing="20.0">
+                 <Label text="Product" styleClass="title" />
+                 <HBox maxWidth="Infinity" spacing="20.0">
                     <VBox.margin>
                        <Insets top="20.0"/>
                     </VBox.margin>
@@ -46,9 +42,6 @@
                     </VBox>
                     <VBox HBox.hgrow="ALWAYS">
                        <HBox spacing="3.0">
-                          <VBox.margin>
-                             <Insets/>
-                          </VBox.margin>
                           <Label styleClass="field-label" text="Category"/>
                           <Label alignment="CENTER" text="*" styleClass="alert"/>
                        </HBox>
@@ -59,7 +52,7 @@
                        </TextField>
                     </VBox>
                  </HBox>
-                 <HBox maxWidth="1.7976931348623157E308" spacing="20.0">
+                 <HBox maxWidth="Infinity" spacing="20.0">
                     <VBox.margin>
                        <Insets top="30.0"/>
                     </VBox.margin>
@@ -84,7 +77,7 @@
                        </TextField>
                     </VBox>
                  </HBox>
-                 <VBox maxHeight="1.7976931348623157E308" VBox.vgrow="ALWAYS">
+                 <VBox maxHeight="Infinity" VBox.vgrow="ALWAYS">
                     <VBox.margin>
                        <Insets top="30.0"/>
                     </VBox.margin>
@@ -94,24 +87,24 @@
                        </VBox.margin>
                        <Label styleClass="field-label" text="Description"/>
                     </HBox>
-                    <TextArea fx:id="description" maxHeight="1.7976931348623157E308"
+                    <TextArea fx:id="description" maxHeight="Infinity"
                               promptText="The best noodle in the world..."
-                              prefRowCount="3"
-                              VBox.vgrow="ALWAYS">
+                              prefRowCount="3" VBox.vgrow="ALWAYS">
                        <VBox.margin>
                           <Insets top="5.0"/>
                        </VBox.margin>
                     </TextArea>
                  </VBox>
-                 <HBox alignment="CENTER_RIGHT" maxWidth="1.7976931348623157E308" spacing="20">
+                 <HBox alignment="CENTER_RIGHT" maxWidth="Infinity" spacing="20">
                     <VBox.margin>
                        <Insets top="20.0" bottom="10.0"/>
                     </VBox.margin>
-                    <Text fx:id="error" wrappingWidth="465" fill="#F84748" />
-                    <Button onMouseClicked="#handleCreateProduct" alignment="CENTER_RIGHT" mnemonicParsing="false"
-                            styleClass="primary-button" text="Add">
+                    <Text fx:id="error" wrappingWidth="465" fill="#F84748"/>
+                    <Button onMouseClicked="#handleCreateProduct" alignment="CENTER_RIGHT"
+                            mnemonicParsing="false" styleClass="primary-button" text="Add">
                        <graphic>
-                          <ImageView fitHeight="14.0" fitWidth="14.0" pickOnBounds="true" preserveRatio="true">
+                          <ImageView fitHeight="14.0" fitWidth="14.0"
+                                     pickOnBounds="true" preserveRatio="true">
                              <Image url="@../../images/plus-circle.png"/>
                           </ImageView>
                        </graphic>
